Reorganize input and output types
