--- TeacherWorkout.Api/GraphQL/Inputs/LessonSaveInput.cs
@@ -1,13 +1,0 @@
-using GraphQL.Types;
-
-namespace TeacherWorkout.Api.GraphQL.Inputs
-{
-    public class LessonSaveInput : InputObjectGraphType
-    {
-        public LessonSaveInput()
-        {
-            Name = "LessonSaveInput";
-            Field<IdGraphType>("lessonId");
-        }
-    }
-}
