+++ TeacherWorkout.Api/Models/Inputs/LessonSaveInput.cs
@@ -1,0 +1,7 @@
+namespace TeacherWorkout.Api.Models.Inputs
+{
+    public class LessonSaveInput
+    {
+        public string LessonId { get; set; }
+    }
+}
